==> code/hospitalisation_death/03_analysis.Rmd <==
---
title: All virally sequenced cases from specimen collection date of first confirmed
  AY4.2 case (8/6/21) - Hospitalisation and Mortality Risks ADULTS
author: "Steven Kerr, Chris Robertson"
date: "27/10/2021"
output:
  html_document:
    df_print: paged
  pdf_document: default
  word_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#Libraries
library(plyr)
library(tidyverse)
library(survival)
library(finalfit)
library(knitr)
library(rmarkdown)
library(flextable)

Location = '/conf/'

setwd(paste0(Location, 'EAVE/GPanalysis/analyses/AY4.2-variant'))

######### 1 Load data #######

df_pos <- readRDS('./data/df_pos.rds')

df_seq <- readRDS('./data/df_seq.rds')

# Create a combine variant, vaccination status column for pyears calculations
df_seq <- mutate(df_seq, variant_vs = as.factor(paste(variant, vs, sep = '_')))

a_end <- max(c(max(df_seq$admission_date, na.rm=T), max(df_seq$discharge_date, na.rm=T))) 
              
rgs <- colnames(df_seq)[startsWith(colnames(df_seq), "Q")]

summary_tbl_wt_chrt <- read_csv("./output/summary_table_weights_cohort.csv")

summary_tbl_wt_pos <- read_csv("./output/summary_table_weights_positive.csv")

summary_tbl_seq <- read_csv("./output/summary_table_weights_seq.csv")

# Start date is the first AY4.2 case
a_begin <- filter(df_seq, variant == 'AY.4.2') %>%
           pull(specimen_date) %>%
           min()
#a_begin <- as.Date("2020-12-08")

source('./code/hospitalisation_death/00_functions.R')

fun.extract <- function(z.fit) {
  #takes a coxph filt using penalised splines and drops off the ps terms
  #make sure no variable begins ps
  z <- summary(z.fit)
  z <- data.frame(z$conf.int)
  z <- z %>% mutate(names = row.names(z)) %>% 
   filter(!(grepl("^ps", names))) %>% 
   dplyr::relocate(names, .before=1) %>% 
   dplyr::select(-exp..coef.)
  names(z) <- c("names","HR","LCL","UCL")
 z
}

fun_print_hr <- function(z,z_var){
 # z_var <- "vsv2_14+"
  z1 <- filter(z,names==z_var)
  z1[c("HR","LCL","UCL")] <- round(z1[c("HR","LCL","UCL")], 2)
  str_c(z1["HR"], ", 95% CI (", z1["LCL"], ", ", z1["UCL"],")")
}

plot_HR <- function(model_fit, term){
# plots hazard ratios for a single term in a fitted model
  
hr <- termplot(model_fit, term = term, se = T, plot = F)

var <- names(hr)

hr <- hr[[var]]

hr <- mutate(hr, ucl = y + 1.96*se,
                 lcl = y - 1.96*se) %>%
      mutate_at(c('y', 'ucl', 'lcl'), exp)

hr <- do.call(data.frame,lapply(hr, function(x) replace(x, is.infinite(x),NA)))

output <- ggplot(data=hr, aes(x=x, y=y)) + geom_line() +
    geom_ribbon(aes(ymin=lcl, ymax=ucl), linetype=2, alpha=0.1, fill = 'steelblue')  + 
    ylab("Hazard Ratio")

if (var == 'ageYear'){
  output <- output + xlab("Age")
} else if (var == 'days'){
  output <- output + xlab("Days since first specimen collection date")
}

output
}
```

## Introduction

The main analyis in this report is based upon all patients who were virally sequenced in Scotland between the specimen collection date of the first AY4.2 case in Scotland - 8th June 2021 - and the last date of hospital admission or discharge amongst all who were virally sequenced - 25th October 2021.

A covid admission to hospital is taken to be individuals who have a positive covid test within 14 days prior to admission or who tested positive within 2 days of admission. Also included are patients who were already in hospital and then tested positive, more than 2 days post admission.  he latter group are not included in the analysis of the risk of hospital admission from the community with covid as they are already in hospital. 

The varaiant variable takes one of the values: alpha, delta, AY4.2 and other. The delta category consists of all delta cases excluding AY4.2.

The viral sequencing data are linked to the EAVE study data of GP clincial conditions.

Hospital admission is derived from the RAPID database only.  This database records admissions to hospital and for many patients there are repeat admissions within a short space of time, sometimes with missing discharge dates.  The rule for combining repeat admission episodes into putative continuous inpatient stays is the same as for the hospital acquired covid infection report.  Episodes where the discharge date on one is the same or the day before the admission date of the following episode are combined; if the discahrge date is missing and a subsequent admission is within 2 days of the previous one the episodes are combined; if the discharge date is missing and there is a subsequent admission 3 or more days in the future then the discahrge date is set to 2 days after admission.

Vaccination status is defined at the date the of the positive PCR test.  It is coded as unvaccinated (uv), one dose and tested positive within 27 days (v1_0_27), one dose and 28+ days (v1_28+) with 2 dose groups - 0-13 days post second dose (v2_0_13) and 14+ days post second dose (v2_14+). In the descriptive tables, 'vs' is the vaccination status variable.

n_tests is number of positive tests before the cohort start date. n_hh_gp is the number of people in their househld. ave_hh_age is average household age. ur6_2016 is urban-rural classification. lab = lh means test was done in a lighthouse lab. 

Cox proportional hazard models are used to estimate the risk of hospitalisation and covid death following a positive covid test result. 'delta' is the reference category for variant, and 'unvaccinated' is the reference category for vaccination status.


## Descriptives

### Cohort

vacc_type_comb is their latest vaccination status.

```{r table_1, echo=FALSE, warning=FALSE, message = FALSE}

summary_tbl_wt_chrt$Characteristic <- replace_na(summary_tbl_wt_chrt$Characteristic, '') 

mytable(summary_tbl_wt_chrt)
```

### All who tested positive during cohort time period

```{r table_2, echo=FALSE, warning=FALSE, message = FALSE}

summary_tbl_wt_pos$Characteristic <- replace_na(summary_tbl_wt_pos$Characteristic, '') 

mytable(summary_tbl_wt_pos)
```


```{r graph_1, echo=FALSE, warning=FALSE, message = FALSE}


knitr::include_graphics('../../output/pos_tests_by_day.png')


```



### All who have viral sequencing data

```{r table_3, echo=FALSE, warning=FALSE, message = FALSE}

summary_tbl_seq$Characteristic <- replace_na(summary_tbl_seq$Characteristic, '') 

mytable(summary_tbl_seq)
```



```{r graph_2, echo=FALSE, warning=FALSE, message = FALSE}
z <- df_seq %>% 
  group_by(specimen_date, variant) %>% 
  dplyr::summarise(N=n()) 

grid <- expand.grid(seq(a_begin, a_end, by="days"), c('delta', 'AY.4.2', 'other'))
        
names(grid) <- c('specimen_date', 'variant')  
       
grid <- grid %>%
       left_join(z) %>%
       replace_na( list(N = 0) )

grid %>%  ggplot(aes(x=specimen_date, y=N, colour = variant)) + geom_point() +
  geom_smooth() +
  labs(x="Specimen Date",y ="Number", title="Number of cases per day", colour="Variant") +         scale_y_log10()
  
# z %>%  ggplot(aes(x=specimen_date, y=N, colour=variant)) + geom_line() +
#   labs(x="Specimen Date",y ="Number", title="Number of cases per day", colour="Variant") + scale_y_log10()
```

### All who were sequenced, stratified by whether they were in hospital at specimen date 


```{r table_4, echo=FALSE, warning=FALSE, message = FALSE}

explanatory <- c("Sex", 
                 "ageYear", 
                 "age_grp",
                 "vs",
                 "lab",
                 "hosp_covid",
                 "hosp_covid_emerg",
                 "death_covid",
                 "death",
                 "simd", 
                 "ur6_2016", 
                 "n_risk_gps",
                 "n_tests", 
                 "ave_hh_age", 
                 "n_hh_gp", 
                 "bmi_cat", 
                 rgs,
                 'EAVE_Smoke',
                 'EAVE_BP')

summary_tbl_seq_in_hosp <- summary_factorlist(df_seq %>% mutate(in_hosp_at_test = as.factor(in_hosp_at_test)), "in_hosp_at_test", explanatory = explanatory, add_col_totals = TRUE)

names(summary_tbl_seq_in_hosp) <- c('Characteristic', 'Levels', 'Not in hospital', 'In hospital')

mytable(summary_tbl_seq_in_hosp)

```

The above tables just describe the cases who test positive after admission - not all will be hospital acquired as the test is 3+ days post admission. 


##Time to emergency covid hospital admission from community

### Number of events and person years of follow up for emergency covid hospital admission from community
```{r analysis_0_pyears, echo=FALSE, warning=FALSE, message = FALSE}

z.rv <- "hosp_covid_emerg" 
z.rv.time <- "time_to_hosp" 
fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```

```{r analysis_0_pyears2, echo=FALSE, warning=FALSE, message = FALSE}

z.rv <- "hosp_covid_emerg" 
z.rv.time <- "time_to_hosp" 

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant_vs "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```

```{r graph_4, echo=FALSE, warning=FALSE, message = FALSE}
z <- df_seq %>% filter(hosp_covid_emerg==1 & in_hosp_at_test == 0 & lab == 'lh') %>% 
  group_by(admission_date, variant) %>% 
  dplyr::summarise(N=n()) %>% 
  filter(admission_date > a_begin)

grid <- expand.grid(seq(a_begin, a_end, by="days"), c('delta', 'AY.4.2', 'other'))
        
names(grid) <- c('admission_date', 'variant')  
       
grid <- grid %>%
       left_join(z) %>%
       replace_na( list(N = 0) )

grid %>%  ggplot(aes(x=admission_date, y=N, colour = variant)) + geom_point() +
  scale_y_continuous(limits = c(0, 12), breaks = 0:max(z$N + 1)) +
  geom_smooth() +
  labs(x="Admission Date",y ="Number", colour="Variant", title="Emergency Covid hospital admissions by day")

# 
# z %>%  ggplot(aes(x=admission_date, y=N, colour = variant)) + geom_point() +
#   scale_y_continuous(limits = c(0, 12), breaks = 0:max(z$N + 1)) +
#   geom_smooth() +
#   labs(x="Admission Date",y ="Number", colour="Variant", title="Emergency Covid hospital admissions by day")
```


### Number of emergency covid hospital admission from community by vaccination status at specimen date

```{r analysis_0_n, echo=FALSE, warning=FALSE, message = FALSE}

df_seq.fit <- df_seq %>% filter(in_hosp_at_test == 0 & lab == 'lh')

z <- df_seq.fit %>% group_by(vs) %>% dplyr::summarise(N=n(), hosp_covid_emerg =sum(hosp_covid_emerg))

mytable(z)

```

### Cumulative risk of emergency covid hospital admission from community by variant

```{r analysis_0_risk, echo=FALSE, warning=FALSE, message = FALSE}
df_seq$variant <- as.factor(df_seq$variant)

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.survfit <- survfit(fmla.plot, data=df_seq, subset= in_hosp_at_test == 0 & lab == 'lh')

plot(z.survfit, fun="event", col=1:4, xlab="Days from Test to emergency hospital admission",
     ylab="Risk")
legend("topleft",col=1:4, lty=1, legend=levels(df_seq$variant) )

```

### Model selection statistics

```{r analysis_0_selection, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) +pspline(days) + Sex + simd + variant +  n_risk_gps +vs "))

z.fit <- coxph(fmla.final , data=df_seq %>% 
                 select(time_to_hosp, hosp_covid_emerg, ageYear, days, Sex, simd,
                        variant, n_risk_gps, vs, in_hosp_at_test, lab) %>% 
                 na.omit(), 
               subset = in_hosp_at_test == 0 & lab == 'lh')

drop1(z.fit, test="Chisq")

```

### Hazard Ratios for emergency covid hospital admission from community

```{r analysis_0_HR, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'lh')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)
    

```

### Hazard Ratios for emergency covid hospital admission from NHS lab

```{r analysis_0_HR_nhs, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'nhs')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)
    

```

### Hazard Ratios for emergency admission from community, with interaction between variant and vaccine status

```{r analysis_0_HR_int, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days) + Sex + simd + n_risk_gps + variant +  variant:vs  "))

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'lh' )

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```

The NAs in the table above are due to the fact that these categories have zero events. 

```{r investigations, eval = TRUE, echo = TRUE, results = 'asis'}
filter(df_seq, variant == 'AY.4.2', vs =='v1_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v1_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v1_28+', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'AY.4.2', vs =='v2_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v2_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v2_28+', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

```



##Time to covid death

### Number of events and person years of follow up for covid deaths

```{r analysis_1_pyears, echo=FALSE, warning=FALSE, message = FALSE}

z.rv <- "death_covid" 
z.rv.time <- "time_to_death" 

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```

```{r analysis_1_pyears2, echo=FALSE, warning=FALSE, message = FALSE}

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant_vs "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```


```{r analysis_1_event_hist, echo=FALSE, warning=FALSE, message = FALSE}
z <- df_seq %>% filter(death_covid==1 & in_hosp_at_test == 0 & lab == 'lh') %>% 
  group_by(admission_date, variant) %>% 
  dplyr::summarise(N=n()) %>% 
  filter(admission_date > a_begin)

grid <- expand.grid(seq(a_begin, a_end, by="days"), c('delta', 'AY.4.2', 'other'))
        
names(grid) <- c('admission_date', 'variant')  
       
grid <- grid %>%
       left_join(z) %>%
       replace_na( list(N = 0) )

grid %>%  ggplot(aes(x=admission_date, y=N, colour = variant)) + geom_point() +
  scale_y_continuous(limits = c(0, 3), breaks = 0:max(z$N + 1)) +
  labs(x="Admission Date",y ="Number", colour="Variant", title="Covid deaths by day")

```


### Number of covid deaths by vaccination status at specimen date


```{r analysis_1_n, echo=FALSE, warning=FALSE, message = FALSE}
df_seq.fit <- df_seq %>% filter(in_hosp_at_test == 0 & lab == 'lh')

z <- df_seq.fit %>% group_by(vs) %>% dplyr::summarise(N=n(), Died=sum(death_covid))

mytable(z)

```

### Cumulative risk of covid death by variant

```{r analysis_1_risk, echo=FALSE, warning=FALSE, message = FALSE}

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.survfit <- survfit(fmla.plot, data=df_seq, subset= in_hosp_at_test == 0 & lab == 'lh')

plot(z.survfit, fun="event", col=1:4, xlab="Days from Test to death",
     ylab="Risk")
legend("topleft",col=1:4, lty=1, legend=levels(df_seq$variant) )

```



### Model selection statistics

```{r analysis_1_selection, echo=FALSE, warning=FALSE, message = FALSE}
fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days) + Sex + simd + variant + n_risk_gps +vs "))

z.fit <- coxph(fmla.final , data=df_seq %>% 
                 select(time_to_death, death_covid, ageYear, days, Sex, simd, variant, n_risk_gps, vs, in_hosp_at_test, lab) %>% 
                 na.omit(), 
               subset = in_hosp_at_test == 0 & lab == 'lh')

drop1(z.fit, test="Chisq")

```


### Hazard Ratios for covid death from community

```{r analysis_1_HR, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'lh')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```


### Hazard Ratios for covid death from NHS lab

```{r analysis_1_HR_nhs, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'nhs')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```


### Hazard Ratios for covid death, with interaction between variant and vaccine status

```{r analysis_1_HR_int, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd + n_risk_gps + variant + variant:vs "))

z.fit <- coxph(fmla.final , data=df_seq.fit, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```

The Infs in the table above are due to the fact that these categories have zero events

```{r investigations_2, eval = TRUE, echo = TRUE, results = 'asis'}

filter(df_seq, variant == 'other', vs =='v2_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(death_covid) %>%
  table()

```




##Time to emergency covid hospitalisation or covid death

### Number of events and person years of follow up for emergency covid hospitalisation or covid death

```{r analysis_2_pyears, echo=FALSE, warning=FALSE, message = FALSE}

z.rv <- "hosp_death_covid" 
z.rv.time <- "time_to_hosp_death" 

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```

```{r analysis_2_pyears2, echo=FALSE, warning=FALSE, message = FALSE}

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant_vs "))
z.tab <- pyears(fmla.plot, data=df_seq, in_hosp_at_test == 0 & lab == 'lh', data.frame=TRUE)$data

mytable(z.tab)

```


```{r analysis_2_event_hist, echo=FALSE, warning=FALSE, message = FALSE}
z <- df_seq %>% filter(hosp_death_covid==1 & in_hosp_at_test == 0 & lab == 'lh') %>%
  mutate(event_date = pmin(NRS.Date.Death, admission_date, na.rm = TRUE) ) %>%
  group_by(event_date, variant) %>% 
  dplyr::summarise(N=n())  


grid <- expand.grid(seq(a_begin, a_end, by="days"), c('delta', 'AY.4.2', 'other'))
        
names(grid) <- c('event_date', 'variant')  
       
grid <- grid %>%
       left_join(z) %>%
       replace_na( list(N = 0) )

grid %>%  ggplot(aes(x=event_date, y=N, colour = variant)) + geom_point() +
  scale_y_continuous(limits = c(0, 12), breaks = 0:max(z$N + 1)) +
  geom_smooth() +
  labs(x="Event date",y ="Number", colour="Variant", title="Emergency covid hospital admissions or covid deaths by day")

```

### Number of emergency covid hospital admissions or covid deaths by vaccination status at specimen date


```{r analysis_2_n, echo=FALSE, warning=FALSE, message = FALSE}
df_seq.fit <- df_seq %>% filter(in_hosp_at_test == 0 & lab == 'lh')

z <- df_seq.fit %>% group_by(vs) %>% dplyr::summarise(N=n(), Died=sum(hosp_death_covid))

mytable(z)

```

### Number of covid deaths that were not covid hospitalisations by variant

```{r death_not_hosp, echo=FALSE, warning=FALSE, message = FALSE}
filter(df_seq, death_covid == 1 & hosp_covid == 0 & in_hosp_at_test == 0 & lab == 'lh') %>%
  group_by(variant) %>%
  summarise(n = n()) %>%
  mytable()
```

### Cumulative risk of emergency covid hospital admission or covid deaths by variant

```{r analysis_2_risk, echo=FALSE, warning=FALSE, message = FALSE}

fmla.plot <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~  variant "))
z.survfit <- survfit(fmla.plot, data=df_seq, subset= in_hosp_at_test == 0 & lab == 'lh')

plot(z.survfit, fun="event", col=1:4, xlab="Days from Test to covid death or emergency covid hospital admission",
     ylab="Risk")
legend("topleft",col=1:4, lty=1, legend=levels(df_seq$variant) )

```

### Model selection statistics

```{r analysis_2_selection, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days) + Sex + simd + n_risk_gps + variant +vs "))

z.fit <- coxph(fmla.final , data=df_seq %>% 
                 select(time_to_hosp_death, hosp_death_covid, ageYear, days, Sex, simd, variant, n_risk_gps, vs, in_hosp_at_test, lab) %>% 
                 na.omit(), 
               subset = in_hosp_at_test == 0 & lab == 'lh')

drop1(z.fit, test="Chisq")

```


### Hazard Ratios for emergency covid hospitalisation or covid death from community

```{r analysis_2_HR, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'lh')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```

### Hazard Ratios for emergency covid hospitalisation or covid death from NHS lab

```{r analysis_2_HR_nhs, echo=FALSE, warning=FALSE, message = FALSE}

z.fit <- coxph(fmla.final , data=df_seq, subset = in_hosp_at_test == 0 & lab == 'nhs')

z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```


#### Hazard Ratios for emergency covid hospitalisation or covid death from community testing, with interaction between variant and vaccine status

```{r analysis_2_HR_int, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + variant + variant:vs "))

z.fit <- coxph(fmla.final , data=df_seq.fit, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)



fmla.test <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + variant + variant*vs "))

z.fit <- coxph(fmla.test , data=df_seq.fit, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

drop1(z.fit, scope ~ variant:vs, test="Chisq")


```






The NA in the table above is due to the fact that the category has zero events

```{r investigations_3, eval = TRUE, echo = TRUE, results = 'asis'}

filter(df_seq, variant == 'AY.4.2', vs =='v1_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'AY.4.2', vs =='v2_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v2_0:27', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()

filter(df_seq, variant == 'other', vs =='v2_28+', in_hosp_at_test == 0, lab == 'lh') %>%
  select(hosp_covid_emerg) %>%
  table()
```



#### Hazard Ratios for emergency covid hospitalisation or covid death from community testing, with interaction between variant and vaccine status by vaccine type

```{r analysis_2_HR_int_type, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + variant + variant:vs_type "))

z.fit <- coxph(fmla.final , data=df_seq.fit, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```


```{r analysis_2_HR_int_type_notMO, echo=FALSE, warning=FALSE, message = FALSE}
df_seq.fit <- df_seq %>% filter(in_hosp_at_test == 0 & lab == 'lh')

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + var + var:vs_t "))

z_df <- df_seq.fit %>% 
        filter( is.na(vacc_type) | vacc_type %in% c('AZ', 'PB') ) %>% 
        filter( is.na(vacc_type_2) | vacc_type_2 %in% c('AZ', 'PB')) %>%
        dplyr::rename(vs_t = vs_type,
                     var = variant)

z.fit <- coxph(fmla.final , data = z_df, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```




#### Time to covid emergency hospitalisation or death amongst all who tested positive, using the date of most recent positive test

This analysis uses everyone who tested positive, rather than only those who were sequenced.

```{r analysis_2_HR_int_pos, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + vs "))

z.fit <- coxph(fmla.final , data=df_pos, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```



```{r analysis_2_HR_int_type_pos, echo=FALSE, warning=FALSE, message = FALSE}

fmla.final <- as.formula(paste("Surv(",z.rv.time,",",z.rv,") ~   pspline(ageYear) + pspline(days)  + Sex  + simd +  n_risk_gps + vs_type "))

z.fit <- coxph(fmla.final , data=df_pos, subset = in_hosp_at_test == 0 & lab == 'lh')
z <- fun.extract(z.fit)

mytable(z)

# Plot HRs for spline terms
plot_HR(z.fit, 1)

plot_HR(z.fit, 2)

```
